requirements.txt
Flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
openai==1.12.0
bcrypt==4.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9 
httpx==0.28.1